Crate three_d

A 3D renderer which enables out-of-the-box build to both desktop and web with the same code. See the README for more information and the examples for how to use it.

Re-exports

pub use renderer::*;

pub use window::*;

Modules

context

Low-level graphics abstraction layer which maps one-to-one with the OpenGL graphics API on native and WebGL2 graphics API on web. This is just a re-export of the glow crate. Use this if you want to have complete control of a feature but be aware that there are no safety checks.

core

Mid-level modular abstractions of common graphics concepts such as buffer, texture, program, render target and so on. Can be combined with low-level calls in the context module as well as high-level functionality in the renderer module.

renderer

High-level features for easy rendering of different types of objects with different types of shading. Can be combined seamlessly with the mid-level features in the core module as well as functionality in the context module.

window

Window, event handling and context creation for easy setup.