Module three_d::core::math

Basic math functionality. Mostly just an re-export of cgmath.

Structs

f16

A 16-bit floating point type implementing the IEEE 754-2008 standard binary16 a.k.a half format.

Traits

Angle

Angles and their associated trigonometric functions.

Array

An array containing elements of type Element

Bounded

Numbers which have upper and lower bounds

ElementWise

Element-wise arithmetic operations. These are supplied for pragmatic reasons, but will usually fall outside of traditional algebraic properties.

EuclideanSpace

Points in a Euclidean space with an associated space of displacement vectors.

InnerSpace

Vectors that also have a dot (or inner) product.

Matrix

A column-major matrix of arbitrary dimensions.

MetricSpace

A type with a distance function between values.

One

Defines a multiplicative identity element for Self.

Rotation

A trait for a generic rotation. A rotation is a transformation that creates a circular motion, and preserves at least one point in the space.

Rotation2

A two-dimensional rotation.

Rotation3

A three-dimensional rotation.

SquareMatrix

A column-major major matrix where the rows and column vectors are of the same dimensions.

Transform

A trait representing an affine transformation that can be applied to points or vectors. An affine transformation is one which

Transform2

Transform3

VectorSpace

Vectors that can be added together and multiplied by scalars.

Zero

Defines an additive identity element for Self.

Functions

degrees

radians

rotation_matrix_from_dir_to_dir

vec2

vec3

vec4

Type Definitions

Degrees

Mat2

Mat3

Mat4

Point

Quat

Radians

Vec2

Vec3

Vec4