Struct three_d::core::prelude::Matrix3

#[repr(C)]
pub struct Matrix3<S> {
    pub x: Vector3<S>,
    pub y: Vector3<S>,
    pub z: Vector3<S>,
}

A 3 x 3, column major matrix

This type is marked as #[repr(C)].

Fields
x: Vector3<S>

The first column of the matrix.

y: Vector3<S>

The second column of the matrix.

z: Vector3<S>

The third column of the matrix.

Implementations

impl<S> Matrix3<S>

pub const fn new( c0r0: S, c0r1: S, c0r2: S, c1r0: S, c1r1: S, c1r2: S, c2r0: S, c2r1: S, c2r2: S ) -> Matrix3<S>

Create a new matrix, providing values for each index.

pub const fn from_cols( c0: Vector3<S>, c1: Vector3<S>, c2: Vector3<S> ) -> Matrix3<S>

Create a new matrix, providing columns.

impl<S> Matrix3<S>
where S: BaseFloat,

pub fn from_translation(v: Vector2<S>) -> Matrix3<S>

Create a homogeneous transformation matrix from a translation vector.

pub fn from_scale(value: S) -> Matrix3<S>

Create a homogeneous transformation matrix from a scale value.

pub fn from_nonuniform_scale(x: S, y: S) -> Matrix3<S>

Create a homogeneous transformation matrix from a set of scale values.

pub fn look_at(dir: Vector3<S>, up: Vector3<S>) -> Matrix3<S>

๐Deprecated: Use Matrix3::look_to_lh

Create a rotation matrix that will cause a vector to point at dir, using up for orientation.

pub fn look_to_lh(dir: Vector3<S>, up: Vector3<S>) -> Matrix3<S>

Create a rotation matrix that will cause a vector to point at dir, using up for orientation.

pub fn look_to_rh(dir: Vector3<S>, up: Vector3<S>) -> Matrix3<S>

Create a rotation matrix that will cause a vector to point at dir, using up for orientation.

pub fn from_angle_x<A>(theta: A) -> Matrix3<S>
where A: Into<Rad<S>>,

Create a rotation matrix from a rotation around the x axis (pitch).

pub fn from_angle_y<A>(theta: A) -> Matrix3<S>
where A: Into<Rad<S>>,

Create a rotation matrix from a rotation around the y axis (yaw).

pub fn from_angle_z<A>(theta: A) -> Matrix3<S>
where A: Into<Rad<S>>,

Create a rotation matrix from a rotation around the z axis (roll).

pub fn from_axis_angle<A>(axis: Vector3<S>, angle: A) -> Matrix3<S>
where A: Into<Rad<S>>,

Create a rotation matrix from an angle around an arbitrary axis.

The specified axis must be normalized, or it represents an invalid rotation.

pub fn is_finite(&self) -> bool

Are all entries in the matrix finite.

impl<S> Matrix3<S>
where S: NumCast + Copy,

pub fn cast<T>(&self) -> Option<Matrix3<T>>
where T: NumCast,

Component-wise casting to another type