Struct three_d::core::prelude::Quaternion

#[repr(C)]
pub struct Quaternion<S> {
    pub v: Vector3<S>,
    pub s: S,
}

A quaternion in scalar/vector form.

This type is marked as #[repr(C)].

Fields
v: Vector3<S>

The vector part of the quaternion.

s: S

The scalar part of the quaternion.

Implementations

impl<S> Quaternion<S>

pub const fn new(w: S, xi: S, yj: S, zk: S) -> Quaternion<S>

Construct a new quaternion from one scalar component and three imaginary components.

pub const fn from_sv(s: S, v: Vector3<S>) -> Quaternion<S>

Construct a new quaternion from a scalar and a vector.

impl<S> Quaternion<S>where S: BaseFloat,

pub fn from_arc( src: Vector3<S>, dst: Vector3<S>, fallback: Option<Vector3<S>> ) -> Quaternion<S>

Construct a new quaternion as a closest arc between two vectors

Return the closest rotation that turns src vector into dst.

[Related StackOverflow question] (http://stackoverflow.com/questions/1171849/finding-quaternion-representing-the-rotation-from-one-vector-to-another)

[Ogre implementation for normalized vectors] (https://bitbucket.org/sinbad/ogre/src/9db75e3ba05c/OgreMain/include/OgreVector3.h?fileviewer=file-view-default#cl-651)

pub fn conjugate(self) -> Quaternion<S>

The conjugate of the quaternion.

pub fn nlerp(self, other: Quaternion<S>, amount: S) -> Quaternion<S>

Do a normalized linear interpolation with other, by amount.

This takes the shortest path, so if the quaternions have a negative dot product, the interpolation will be between self and -other.

pub fn slerp(self, other: Quaternion<S>, amount: S) -> Quaternion<S>

Spherical Linear Interpolation

Return the spherical linear interpolation between the quaternion and other. Both quaternions should be normalized first.

This takes the shortest path, so if the quaternions have a negative dot product, the interpolation will be between self and -other.

Performance notes

The acos operation used in slerp is an expensive operation, so unless your quaternions are far away from each other it’s generally more advisable to use nlerp when you know your rotations are going to be small.

[Understanding Slerp, Then Not Using It] (http://number-none.com/product/Understanding%20Slerp,%20Then%20Not%20Using%20It/)

[Arcsynthesis OpenGL tutorial] (http://www.arcsynthesis.org/gltut/Positioning/Tut08%20Interpolation.html)

pub fn is_finite(&self) -> bool

impl<S> Quaternion<S>where S: NumCast + Copy,

pub fn cast<T>(&self) -> Option<Quaternion<T>>where T: BaseFloat,

Component-wise casting to another type.