Struct three_d::core::render_target::ColorTarget

pub struct ColorTarget<'a> { /* private fields */ }

Adds additional functionality to clear, read from and write to a texture. Use the as_color_target function directly on the texture structs (for example Texture2D) to construct a color target. Combine this together with a DepthTarget with RenderTarget::new to be able to write to both a depth and color target at the same time. A color target purely adds functionality, so it can be created each time it is needed, the actual data is saved in the texture.

Note: DepthTest is disabled if not also writing to a DepthTarget.

Implementations

impl<'a> ColorTarget<'a>

pub fn clear(&self, clear_state: ClearState) -> &Self

Clears the color of this color target as defined by the given clear state.

pub fn clear_partially( &self, scissor_box: ScissorBox, clear_state: ClearState ) -> &Self

Clears the color of the part of this color target that is inside the given scissor box.

pub fn write(&self, render: impl FnOnce()) -> &Self

Writes whatever rendered in the render closure into this color target.

pub fn write_partially( &self, scissor_box: ScissorBox, render: impl FnOnce() ) -> &Self

Writes whatever rendered in the render closure into the part of this color target defined by the scissor box.

pub fn read<T: TextureDataType>(&self) -> Vec<T>

Returns the colors of the pixels in this color target. The number of channels per pixel and the data format for each channel is specified by the generic parameter.

Note: On web, the data format needs to match the data format of the color texture.

pub fn read_partially<T: TextureDataType>( &self, scissor_box: ScissorBox ) -> Vec<T>

Returns the colors of the pixels in this color target inside the given scissor box. The number of channels per pixel and the data format for each channel is specified by the generic parameter.

Note: On web, the data format needs to match the data format of the color texture.

pub fn copy_from( &self, color_texture: ColorTexture<'_>, viewport: Viewport, write_mask: WriteMask ) -> &Self

👎Deprecated: use apply_screen_effect with a CopyEffect instead

Copies the content of the color texture as limited by the WriteMask to the part of this color target specified by the Viewport.

pub fn copy_partially_from( &self, scissor_box: ScissorBox, color_texture: ColorTexture<'_>, viewport: Viewport, write_mask: WriteMask ) -> &Self

👎Deprecated: use apply_screen_effect_partially with a CopyEffect instead

Copies the content of the color texture as limited by the ScissorBox and WriteMask to the part of this color target specified by the Viewport.

pub fn width(&self) -> u32

Returns the width of the color target in texels. If using the zero mip level of the underlying texture, then this is simply the width of that texture, otherwise it is the width of the given mip level.

pub fn height(&self) -> u32

Returns the height of the color target in texels. If using the zero mip level of the underlying texture, then this is simply the height of that texture, otherwise it is the height of the given mip level.

impl<'a> ColorTarget<'a>

pub fn scissor_box(&self) -> ScissorBox

Returns the scissor box that encloses the entire target.

pub fn viewport(&self) -> Viewport

Returns the viewport that encloses the entire target.

impl<'a> ColorTarget<'a>

pub fn render( &self, camera: &Camera, objects: impl IntoIterator<Item = impl Object>, lights: &[&dyn Light] ) -> &Self

Render the objects using the given camera and lights into this render target. Use an empty array for the lights argument, if the objects does not require lights to be rendered. Also, objects outside the camera frustum are not rendered and the objects are rendered in the order given by cmp_render_order.

pub fn render_partially( &self, scissor_box: ScissorBox, camera: &Camera, objects: impl IntoIterator<Item = impl Object>, lights: &[&dyn Light] ) -> &Self

Render the objects using the given camera and lights into the part of this render target defined by the scissor box. Use an empty array for the lights argument, if the objects does not require lights to be rendered. Also, objects outside the camera frustum are not rendered and the objects are rendered in the order given by cmp_render_order.

pub fn render_with_material( &self, material: &dyn Material, camera: &Camera, geometries: impl IntoIterator<Item = impl Geometry>, lights: &[&dyn Light] ) -> &Self

Render the geometries with the given Material using the given camera and lights into this render target. Use an empty array for the lights argument, if the material does not require lights to be rendered.

pub fn render_partially_with_material( &self, scissor_box: ScissorBox, material: &dyn Material, camera: &Camera, geometries: impl IntoIterator<Item = impl Geometry>, lights: &[&dyn Light] ) -> &Self

Render the geometries with the given Material using the given camera and lights into the part of this render target defined by the scissor box. Use an empty array for the lights argument, if the material does not require lights to be rendered.

pub fn render_with_effect( &self, effect: &dyn Effect, camera: &Camera, geometries: impl IntoIterator<Item = impl Geometry>, lights: &[&dyn Light], color_texture: Option<ColorTexture<'_>>, depth_texture: Option<DepthTexture<'_>> ) -> &Self

Render the geometries with the given Effect using the given camera and lights into this render target. Use an empty array for the lights argument, if the effect does not require lights to be rendered.

pub fn render_partially_with_effect( &self, scissor_box: ScissorBox, effect: &dyn Effect, camera: &Camera, geometries: impl IntoIterator<Item = impl Geometry>, lights: &[&dyn Light], color_texture: Option<ColorTexture<'_>>, depth_texture: Option<DepthTexture<'_>> ) -> &Self

Render the geometries with the given Effect using the given camera and lights into the part of this render target defined by the scissor box. Use an empty array for the lights argument, if the effect does not require lights to be rendered.

pub fn apply_screen_material( &self, material: &dyn Material, camera: &Camera, lights: &[&dyn Light] ) -> &Self

Apply the given Material to this render target. Use an empty array for the lights argument, if the material does not require lights to be rendered.

pub fn apply_screen_material_partially( &self, scissor_box: ScissorBox, material: &dyn Material, camera: &Camera, lights: &[&dyn Light] ) -> &Self

Apply the given Material to the part of this render target defined by the scissor box. Use an empty array for the lights argument, if the material does not require lights to be rendered.

pub fn apply_screen_effect( &self, effect: &dyn Effect, camera: &Camera, lights: &[&dyn Light], color_texture: Option<ColorTexture<'_>>, depth_texture: Option<DepthTexture<'_>> ) -> &Self

Apply the given Effect to this render target. Use an empty array for the lights argument, if the effect does not require lights to be rendered.

pub fn apply_screen_effect_partially( &self, scissor_box: ScissorBox, effect: &dyn Effect, camera: &Camera, lights: &[&dyn Light], color_texture: Option<ColorTexture<'_>>, depth_texture: Option<DepthTexture<'_>> ) -> &Self

Apply the given Effect to the part of this render target defined by the scissor box. Use an empty array for the lights argument, if the effect does not require lights to be rendered.