Struct three_d::core::AxisAlignedBoundingBox

pub struct AxisAlignedBoundingBox { /* fields omitted */ }

A bounding box that aligns with the x, y and z axes.

Implementations

impl AxisAlignedBoundingBox

pub const EMPTY: Self

An empty bounding box.

pub const INFINITE: Self

An infinitely large bounding box.

pub fn new_with_positions(positions: &[f32]) -> Self

Constructs a new bounding box and expands it such that all of the given positions are contained inside the bounding box. A position consisting of an x, y and z coordinate corresponds to three consecutive value in the positions array.

pub fn new_with_transformed_positions(
positions: &[f32],
transformation: &Mat4
) -> Self

Constructs a new bounding box and expands it such that all of the given positions transformed with the given transformation are contained inside the bounding box. A position consisting of an x, y and z coordinate corresponds to three consecutive value in the positions array.

pub fn is_empty(&self) -> bool

Returns true if the bounding box is empty (ie. constructed by AxisAlignedBoundingBox::EMPTY).

pub fn is_infinite(&self) -> bool

Returns true if the bounding box is infinitely large (ie. constructed by AxisAlignedBoundingBox::INFINITE).

pub fn min(&self) -> Vec3

Get the minimum coordinate of the bounding box.

pub fn max(&self) -> Vec3

Get the maximum coordinate of the bounding box.

pub fn center(&self) -> Vec3

Get the center of the bounding box.

pub fn size(&self) -> Vec3

Get the size of the bounding box.

pub fn expand(&mut self, positions: &[f32])

Expands the bounding box such that all of the given positions are contained inside the bounding box. A position consisting of an x, y and z coordinate corresponds to three consecutive value in the positions array.

pub fn expand_with_transformation(
&mut self,
positions: &[f32],
transformation: &Mat4
)

Expands the bounding box such that all of the given positions transformed with the given transformation are contained inside the bounding box. A position consisting of an x, y and z coordinate corresponds to three consecutive value in the positions array.

pub fn expand_with_aabb(&mut self, other: &AxisAlignedBoundingBox)

Expand the bounding box such that it also contains the given other bounding box.

pub fn transform(&mut self, transformation: &Mat4)

Transforms the bounding box by the given transformation.

Note: Use new_with_transformed_positions instead of new_with_positions followed by this method to create a more tight bounding box.

pub fn distance(&self, position: &Vec3) -> f32

The distance from position to the point in this bounding box that is closest to position.

pub fn distance_max(&self, position: &Vec3) -> f32

The distance from position to the point in this bounding box that is furthest away from position.