Struct three_d::core::ImageEffect

pub struct ImageEffect { /* fields omitted */ }

A customizable 2D effect. Can for example be used for adding an effect on top of a rendered image.

Implementations

impl ImageEffect

pub fn new(context: &Context, fragment_shader: &str) -> ThreeDResult<Self>

Creates a new image effect which applies the calculations defined in the given fragment shader source when calling the ImageEffect::apply function.

pub fn render(
&self,
render_states: RenderStates,
viewport: Viewport
) -> ThreeDResult<()>

Applies the calculations defined in the fragment shader given at construction and output it to the current screen/render target. Must be called in a render target render function, for example in the callback function of Screen::write.

pub fn apply(
&self,
render_states: RenderStates,
viewport: Viewport
) -> ThreeDResult<()>

Applies the calculations defined in the fragment shader given at construction and output it to the current screen/render target. Must be called in a render target render function, for example in the callback function of Screen::write.