Module three_d::core::texture

Different types of textures used by the GPU to read from and write to.

Structs

CpuTexture

A CPU-side version of a 2D texture.

CpuTexture3D

A CPU-side version of a 3D texture.

CubeMapSideIterator

Iterator over the 6 side of a cube map.

DepthTexture2D

A 2D depth texture that can be rendered into and read from. See also RenderTarget and DepthTarget.

DepthTexture2DArray

An array of 2D depth textures that can be rendered into and read from. See also RenderTarget and DepthTarget.

DepthTextureCubeMap

A depth texture cube map that can be rendered into and read from. See also RenderTarget and DepthTarget.

Texture2D

A 2D texture, basically an image that is transferred to the GPU.

Texture2DArray

A array of 2D color textures that can be rendered into.

Texture3D

A 3D color texture.

TextureCubeMap

A texture that covers all 6 sides of a cube.

f24

24 bit float which can be used as DepthTextureDataType.

Enums

ColorTexture

A reference to some type of texture containing colors.

CubeMapSide

The 6 sides of a cube map

DepthTexture

A reference to some type of texture containing depths.

Interpolation

Possible modes of interpolation which determines the texture output between texture pixels.

TextureData

The pixel/texel data for a Texture2D or Texture3D.

Wrapping

Possible wrapping modes for a texture which determines how the texture is applied outside of the [0..1] uv coordinate range.

Traits

DepthTextureDataType

The basic data type used for each pixel in a depth texture.

TextureDataType

The basic data type used for each channel of each pixel in a texture.