Trait three_d::core::texture::Texture

pub trait Texture {
    fn bind(&self, location: u32);

    fn width(&self) -> u32;

    fn height(&self) -> u32;

    fn format(&self) -> Format;

    fn is_transparent(&self) -> bool;
}

A texture that can be sampled in a fragment shader (see use_texture).

Required methods

fn bind(&self, location: u32)

Binds this texture to the current shader program.

fn width(&self) -> u32

The width of this texture.

fn height(&self) -> u32

The height of this texture.

fn format(&self) -> Format

The format of this texture.

fn is_transparent(&self) -> bool

Whether this texture contain pixels with alpha value less than maximum.