Trait three_d::core::texture::TextureArray

pub trait TextureArray {
    fn bind(&self, location: u32);

    fn width(&self) -> u32;

    fn height(&self) -> u32;

    fn depth(&self) -> u32;

    fn format(&self) -> Format;
}

A texture array that can be sampled in a fragment shader (see use_texture_array).

Required methods

fn bind(&self, location: u32)

Binds this texture array to the current shader program.

fn width(&self) -> u32

The width of this texture.

fn height(&self) -> u32

The height of this texture.

fn depth(&self) -> u32

The depth of this texture, ie. the number of layers.

fn format(&self) -> Format

The format of this texture.