Trait three_d::core::texture::TextureCube

pub trait TextureCube {
    fn bind(&self, location: u32);

    fn width(&self) -> u32;

    fn height(&self) -> u32;

    fn format(&self) -> Format;

    fn is_hdr(&self) -> bool;
}

A texture cube that can be sampled in a fragment shader (see use_texture_cube).

Required methods

fn bind(&self, location: u32)

Binds this texture cube to the current shader program.

fn width(&self) -> u32

The width of one of the sides of this texture.

fn height(&self) -> u32

The height of one of the sides of this texture.

fn format(&self) -> Format

The format of this texture.

fn is_hdr(&self) -> bool

Whether or not this texture contain hdr information.