Trait three_d::core::UniformDataType

pub trait UniformDataType: DataType { }

Possible types that can be send as a uniform to a shader (a variable that is uniformly available when processing all vertices and fragments).