Function three_d::io::hdr_image_from_bytes

pub fn hdr_image_from_bytes(bytes: &[u8]) -> ThreeDResult<CPUTexture<f32>>

Deserialize the given bytes representing a hdr image into a CPUTexture using the image crate. The CPUTexture can then be used to create a Texture2D or a TextureCubeMap using the new_from_equirectangular method. Supported formats: HDR.