Type Alias three_d::renderer::effect::WaterMaterial

pub type WaterMaterial = WaterEffect;

ğŸ‘Deprecated: renamed to WaterEffect

An effect that simulates a water surface and should therefore only be applied to a water surface geometry. This effect needs the rendered scene (without the water surface) in a color and depth texture to be able to add reflections and refractions.

Aliased Type

struct WaterMaterial {
    pub background: Background,
    pub metallic: f32,
    pub roughness: f32,
    pub lighting_model: LightingModel,
}

Fields
background: Background

The background of the scene which is used for reflections.

metallic: f32

A value in the range [0..1] specifying how metallic the surface is.

roughness: f32

A value in the range [0..1] specifying how rough the surface is.

lighting_model: LightingModel

The lighting model used when rendering this effect