Function three_d::renderer::render_with_material

pub fn render_with_material(
    context: &Context,
    camera: &Camera,
    geometry: impl Geometry,
    material: impl Material,
    lights: &[&dyn Light]
)

Render the given Geometry with the given Material. Must be called in the callback given as input to a RenderTarget, ColorTarget or DepthTarget write method. Use an empty array for the lights argument, if the material does not require lights to be rendered.