Struct three_d::renderer::geometry::Mesh

pub struct Mesh { /* private fields */ }

A triangle mesh Geometry.

Implementations

impl Mesh

pub fn new(context: &Context, cpu_mesh: &CpuMesh) -> Self

Creates a new triangle mesh from the given CpuMesh. All data in the CpuMesh is transfered to the GPU, so make sure to remove all unnecessary data from the CpuMesh before calling this method.

pub fn transformation(&self) -> Mat4

Returns the local to world transformation applied to this mesh.

pub fn set_transformation(&mut self, transformation: Mat4)

Set the local to world transformation applied to this mesh. If any animation method is set using Self::set_animation, the transformation from that method is applied before this transformation.

pub fn set_animation( &mut self, animation: impl Fn(f32) -> Mat4 + Send + Sync + 'static )

Specifies a function which takes a time parameter as input and returns a transformation that should be applied to this mesh at the given time. To actually animate this mesh, call Geometry::animate at each frame which in turn evaluates the animation function defined by this method. This transformation is applied first, then the local to world transformation defined by Self::set_transformation.