Struct three_d::renderer::geometry::ParticleSystem

pub struct ParticleSystem {
    pub acceleration: Vec3,
    /* private fields */
}

Particle system that can be used to simulate effects such as fireworks, fire, smoke or water particles.

All particles are initialised with Particles::start_positions and Particles::start_velocities and a global ParticleSystem::acceleration. Then, when time passes, their position is updated based on

new_position = start_position + start_velocity * time + 0.5 * acceleration * time * time

The particles will only move if the ParticleSystem::animate is called every frame.

Fields
acceleration: Vec3

The acceleration applied to all particles defined in the world coordinate system.

Implementations

impl ParticleSystem

pub fn new( context: &Context, particles: &Particles, acceleration: Vec3, cpu_mesh: &CpuMesh ) -> Self

Creates a new particle system with the given geometry and the given attributes for each particle. The acceleration is applied to all particles defined in the world coordinate system.

pub fn transformation(&self) -> Mat4

Returns local to world transformation applied to the particle geometry before its position is updated as described in ParticleSystem.

pub fn set_transformation(&mut self, transformation: Mat4)

Set the local to world transformation applied to the particle geometry before its position is updated as described in ParticleSystem.

pub fn set_particles(&mut self, particles: &Particles)

Set the particles attributes.