Struct three_d::renderer::light::AmbientLight

pub struct AmbientLight {
    pub intensity: f32,
    pub color: Srgba,
    pub environment: Option<Environment>,
}

A light which shines on all surfaces. Can be uniform (a light that shines equally on any surface) or calculated from an environment map using the Environment struct.

Fields
intensity: f32

The intensity of the light. This allows for higher intensity than 1 which can be used to simulate high intensity light sources like the sun.

color: Srgba

The base color of the light.

environment: Option<Environment>

The light shining from the environment. This is calculated based on an environment map.

Implementations

impl AmbientLight

pub fn new(_context: &Context, intensity: f32, color: Srgba) -> Self

Constructs an ambient light that shines equally on all surfaces.

pub fn new_with_environment( context: &Context, intensity: f32, color: Srgba, environment_map: &TextureCubeMap ) -> Self

Constructs an ambient light that shines based on the given environment map.