Struct three_d::renderer::light::SpotLight

pub struct SpotLight {
    pub intensity: f32,
    pub color: Srgba,
    pub position: Vec3,
    pub direction: Vec3,
    pub cutoff: Radians,
    pub attenuation: Attenuation,
    /* private fields */
}

A light which shines from the given position and in the given direction. The light will cast shadows if you generate a shadow map.

Fields
intensity: f32

The intensity of the light. This allows for higher intensity than 1 which can be used to simulate high intensity light sources like the sun.

color: Srgba

The base color of the light.

position: Vec3

The position of the light.

direction: Vec3

The direction the light shines.

cutoff: Radians

The cutoff angle for the light.

attenuation: Attenuation

The Attenuation of the light.

Implementations

impl SpotLight

pub fn new( context: &Context, intensity: f32, color: Srgba, position: &Vec3, direction: &Vec3, cutoff: impl Into<Radians>, attenuation: Attenuation ) -> SpotLight

Constructs a new spot light.

pub fn clear_shadow_map(&mut self)

Clear the shadow map, effectively disable the shadow. Only necessary if you want to disable the shadow, if you want to update the shadow, just use SpotLight::generate_shadow_map.

pub fn generate_shadow_map( &mut self, texture_size: u32, geometries: impl IntoIterator<Item = impl Geometry> + Clone )

Generate a shadow map which is used to simulate shadows from the spot light onto the geometries given as input. It is recomended that the texture size is power of 2. If the shadows are too low resolution (the edges between shadow and non-shadow are pixelated) try to increase the texture size.

pub fn shadow_map(&self) -> Option<&DepthTexture2D>

Returns a reference to the shadow map if it has been generated.