Struct three_d::renderer::object::Imposters

pub struct Imposters { /* private fields */ }

A level-of-detail technique to replace rendering of high-poly meshes. Should only be used where details cannot be seen, for example when the objects are far away. A set of objects are rendered from different angles into a set of textures and the textures are then rendered continuously instead of the expensive objects.

Implementations

impl Imposters

pub fn new( context: &Context, positions: &[Vec3], objects: impl IntoIterator<Item = impl Object> + Clone, lights: &[&dyn Light], max_texture_size: u32 ) -> Self

Constructs a new Imposters and render the imposter texture from the given objects with the given lights. The imposters are placed at the given positions.

pub fn set_positions(&mut self, positions: &[Vec3])

Set the positions of the imposters.

pub fn update_texture( &mut self, objects: impl IntoIterator<Item = impl Object> + Clone, lights: &[&dyn Light], max_texture_size: u32 )

Render the imposter texture from the given objects with the given lights. Use this if you want to update the look of the imposters.