Struct three_d::renderer::object::InstancedModelPart

pub struct InstancedModelPart<M: Material> { /* private fields */ }

Part of an InstancedModel consisting of a InstancedMesh, some type of material and a set of possible animations.

Implementations

impl<M: Material> InstancedModelPart<M>

pub fn animations(&self) -> Vec<Option<String>>

Returns a list of unique names for the animations for this model part. Use these names as input to Self::choose_animation.

pub fn choose_animation(&mut self, animation_name: Option<&str>)

Specifies the animation to use when Geometry::animate is called. Use the Self::animations method to get a list of possible animations.