Struct three_d::renderer::object::Water

pub struct Water<M: Material> { /* private fields */ }

A water geometry with an applied material.

Implementations

impl<M: Material> Water<M>

pub fn new( context: &Context, material: M, height: f32, center: Vec2, side_length: f32, vertex_distance: f32, parameters: impl IntoIterator<Item = WaveParameters> + Clone ) -> Self

Constructs a new Water object with the given material and at the given height.

The center is the center of the visualized water surface, which can be updated using Self::set_center to simualte an infinite water surface. The side_length is the length of one side of the visualized water surface. The vertex_distance is the distance between vertices.

pub fn set_center(&mut self, center: Vec2)

Set the center of the water. To be able to move the water with the camera, thereby simulating infinite water.

pub fn set_height(&mut self, height: f32)

Set the average height of the water.

pub fn set_parameters( &mut self, parameters: impl IntoIterator<Item = WaveParameters> + Clone )

Set the currently used WaveParameters.

pub fn animate(&mut self, time: f32)

For updating the animation. The time parameter should be some continious time, for example the time since start.