Trait three_d::renderer::object::Object

pub trait Object: Geometry {
    // Required methods
    fn render(&self, camera: &Camera, lights: &[&dyn Light]);
    fn material_type(&self) -> MaterialType;
}

Represents a 3D object which can be rendered directly or used in a render call, for example RenderTarget::render.

Required Methods

fn render(&self, camera: &Camera, lights: &[&dyn Light])

Render the object. Use an empty array for the lights argument, if the objects does not require lights to be rendered. Must be called in the callback given as input to a RenderTarget, ColorTarget or DepthTarget write method.

fn material_type(&self) -> MaterialType

Returns the type of material applied to this object.