Trait three_d::renderer::object::Object2D

pub trait Object2D: Shadable2D {
    fn render(&self, viewport: Viewport) -> ThreeDResult<()>;

    fn is_transparent(&self) -> bool;
}

Represents a 2D object which can be rendered.

Required methods

fn render(&self, viewport: Viewport) -> ThreeDResult<()>

Render the object. Must be called in a render target render function, for example in the callback function of Screen::write.

fn is_transparent(&self) -> bool

Returns whether or not this object should be considered transparent.