Trait three_d::renderer::object::Shadable2D

pub trait Shadable2D {
    fn render_with_material(
        &self, 
        material: &dyn Material, 
        viewport: Viewport
    ) -> ThreeDResult<()>;

    fn render_forward(
        &self, 
        material: &dyn Material, 
        viewport: Viewport
    ) -> ThreeDResult<()>;
}

Represents a 2D object that is possible to render with Materials.

Required methods

fn render_with_material(
&self,
material: &dyn Material,
viewport: Viewport
) -> ThreeDResult<()>

Render the object with the given material. Must be called in a render target render function, for example in the callback function of Screen::write.

fn render_forward(
&self,
material: &dyn Material,
viewport: Viewport
) -> ThreeDResult<()>

👎 Deprecated:
use render_with_material instead

Render the object with the given material. Must be called in a render target render function, for example in the callback function of Screen::write.