Struct three_d::renderer::Camera

pub struct Camera {
    pub tone_mapping: ToneMapping,
    pub color_mapping: ColorMapping,
    /* private fields */
}

Represents a camera used for viewing 2D and 3D objects.

Fields
tone_mapping: ToneMapping

This tone mapping is applied to the final color of renders using this camera.

color_mapping: ColorMapping

This color mapping is applied to the final color of renders using this camera.

Implementations

impl Camera

pub fn new_orthographic( viewport: Viewport, position: Vec3, target: Vec3, up: Vec3, height: f32, z_near: f32, z_far: f32 ) -> Self

New camera which projects the world with an orthographic projection.

pub fn new_perspective( viewport: Viewport, position: Vec3, target: Vec3, up: Vec3, field_of_view_y: impl Into<Radians>, z_near: f32, z_far: f32 ) -> Self

New camera which projects the world with a perspective projection.

pub fn new_2d(viewport: Viewport) -> Self

Returns an orthographic camera for viewing 2D content. The camera is placed at the center of the given viewport. The (0, 0) position is at the bottom left corner and the (viewport.width, viewport.height) position is at the top right corner.

pub fn disable_tone_and_color_mapping(&mut self)

Disables the tone and color mapping so as to be ready for rendering into an intermediate render target with this camera.

pub fn set_default_tone_and_color_mapping(&mut self)

Sets the tone and color mapping to default so as to be ready for rendering into the final render target (usually the screen) with this camera.