Struct three_d::window::WindowedContext

pub struct WindowedContext { /* private fields */ }

A windowed graphics context, ie. a graphics context that is associated with a window. For a graphics context that is not associated with a window, see HeadlessContext.

Implementations

impl WindowedContext

pub fn from_winit_window( window: &Window, settings: SurfaceSettings ) -> Result<Self, WindowError>

Creates a new windowed context from a winit window.

pub fn resize(&self, physical_size: PhysicalSize<u32>)

Resizes the context

pub fn make_current(&self) -> Result<(), WindowError>

Make this context current. Needed when using multiple windows (contexts) on native.

pub fn swap_buffers(&self) -> Result<(), WindowError>

Swap buffers - should always be called after rendering.