Function three_d::core::apply_cube_effect

pub fn apply_cube_effect(
    context: &Context,
    side: CubeMapSide,
    fragment_shader_source: &str,
    render_states: RenderStates,
    viewport: Viewport,
    use_uniforms: impl FnOnce(&Program)
)

👎Deprecated: Use apply_screen_material or apply_screen_effect in the renderer module

Applies a 2D/screen space effect to the given viewport of the given side of a cube map. The fragment shader get the 3D position (specified by in vec3 pos;) of the fragment on the cube with minimum position (-1, -1, -1) and maximum position (1, 1, 1).