Module three_d::core::render_target

Functionality for rendering to the screen or into textures.

Structs

ClearState

Defines which channels (red, green, blue, alpha and depth) to clear when starting to write to a RenderTarget or the Screen. If None then the channel is not cleared and if Some(value) the channel is cleared to that value (the value must be between 0 and 1).

RenderTarget

Adds additional functionality to write to and copy from both a ColorTargetTexture2D and a DepthTargetTexture2D at the same time. It purely adds functionality, so it can be created each time it is needed, the data is saved in the textures.

RenderTargetArray

Adds additional functionality to write to and copy from both a Texture2DArrayand a DepthTargetTexture2DArray at the same time. It purely adds functionality, so it can be created each time it is needed, the data is saved in the textures.

RenderTargetCubeMap

Adds additional functionality to write to a TextureCubeMap and a DepthTargetTextureCubeMap at the same time. It purely adds functionality, so it can be created each time it is needed, the data is saved in the textures.

Screen

The screen render target which is essential to get something on the screen (see the write function).

Enums

CopyDestination

The destination of applying a copy.