Module three_d::core::texture

Different types of textures used by the GPU to read from and write to.

Structs

CPUTexture

A CPU-side version of a Texture2D. Can be constructed manually or loaded via Loader.

CPUTextureCube

A CPU-side version of a TextureCubeMap. All 6 images must have the same dimensions. Can be constructed manually or loaded via Loader.

ColorTargetTexture2DDeprecated

A 2D color texture that can be rendered into and read from.

CubeMapSideIterator

Iterator over the 6 side of a cube map.

DepthTargetTexture2D

A 2D depth texture that can be rendered into and read from. See also RenderTarget.

DepthTargetTexture2DArray

An array of 2D depth textures that can be rendered into and read from. See also RenderTargetArray.

DepthTargetTextureCubeMap

A depth texture cube map that can be rendered into and read from. See also RenderTargetCubeMap.

Texture2D

A 2D texture, basically an image that is transferred to the GPU.

Texture2DArray

A array of 2D color textures that can be rendered into.

TextureCubeMap

A texture that covers all 6 sides of a cube.

Enums

CubeMapSide

The 6 sides of a cube map

DepthFormat

Type of formats for depth render targets (DepthTargetTexture2D and DepthTargetTexture2DArray).

Format

Possible formats for pixels in a texture.

Interpolation

Possible modes of interpolation which determines the texture output between texture pixels.

Wrapping

Possible wrapping modes for a texture which determines how the texture is applied outside of the [0..1] uv coordinate range.

Traits

Texture

A texture that can be sampled in a fragment shader (see use_texture).

TextureArray

A texture array that can be sampled in a fragment shader (see use_texture_array).

TextureCube

A texture cube that can be sampled in a fragment shader (see use_texture_cube).

TextureDataType

The basic data type used for each channel of each pixel in a texture.

Type Definitions

ColorTargetTexture2DArrayDeprecated

A array of 2D color textures that can be rendered into.