Module three_d::renderer::effect

A collection of effects implementing the Effect trait.

Structs

CopyEffect

Copies the content of the color and/or depth texture by rendering a quad with those textures applied. The difference from ScreenEffect is that this effect does not apply any mapping set in the Camera.

FogEffect

An effect that simulates fog, ie. the area where it is applied gets hazy when objects are far away.

FxaaEffect

A simple anti-aliasing approach which smooths otherwise jagged edges (for example lines) but also smooths the rest of the image.

ScreenEffect

Renders a full screen quad with the content of the color and/or depth textures. The difference from CopyEffect is that this effect also applies any mapping set in the Camera.

WaterEffect

An effect that simulates a water surface and should therefore only be applied to a water surface geometry. This effect needs the rendered scene (without the water surface) in a color and depth texture to be able to add reflections and refractions.

Enums

Background

The background of the scene.

Traits

Effect

Similar to Material, the difference is that an effect needs the rendered color texture and/or depth texture of the scene to be applied. Therefore an effect is always applied one at a time and after the scene has been rendered with the regular Material.

Type Aliases

WaterMaterialDeprecated

An effect that simulates a water surface and should therefore only be applied to a water surface geometry. This effect needs the rendered scene (without the water surface) in a color and depth texture to be able to add reflections and refractions.