Module three_d::renderer::material

Contain a collection of common materials to apply to some geometry, for example a Model. It is possible to create a custom material by extending the Material trait.

Structs

CPUMaterial

A CPU-side version of a material. Can be constructed manually or loaded via io.

Color

Represents a color composed of a red, green and blue component. In addition, the alpha value determines the how transparent the color is (0 is fully transparent and 255 is fully opaque).

ColorMaterial

A material that renders a Shadable object in a color defined by multiplying a color with an optional texture and optional per vertex colors. This material is not affected by lights.

DeferredPhysicalMaterial

The deferred part of a physically-based material that renders a Shadable object in an approximate correct physical manner based on Physically Based Rendering (PBR). Must be used together with a DeferredPipeline. This material is affected by lights.

DepthMaterial

NormalMaterial

ORMMaterial

PhysicalMaterial

A physically-based material that renders a Shadable object in an approximate correct physical manner based on Physically Based Rendering (PBR). This material is affected by lights.

PositionMaterial

UVMaterial

Traits

Material

Represents a material that can be applied to a Shadable object.