Module three_d::renderer::object

A collection of objects that can be rendered, for example a mesh.

Structs

Axes

Three arrows indicating the three main axes; the x-axis (red), the y-axis (green) and the z-axis (blue). Used for easily debugging where objects are placed in the 3D world.

AxisAlignedBoundingBox

A bounding box that aligns with the x, y and z axes.

BoundingBox

CPUMesh

A CPU-side version of a triangle mesh. Can be constructed manually or loaded via io or via the utility functions for generating simple triangle meshes.

Circle

Imposters

A level-of-detail technique to replace rendering high-poly meshes at a distance. A mesh is rendered from different angles into a set of textures and the textures are then rendered continuously instead of the high-poly meshes.

InstancedModel

Similar to Model, except it is possible to render many instances of the same model efficiently.

Line

Model

A 3D model consisting of a triangle mesh and any material that implements the Material trait.

ModelInstance

ParticleData

Used to define the initial position and velocity of a particle in Particles.

Particles

Particle effect that can be rendered with any material.

Rectangle

Skybox

An illusion of a sky.

Enums

Indices

An array of indices. Supports different data types.

Traits

Geometry

Represents a 3D geometry.

GeometryMut

Represents a 3D geometry.

Object

Represents a 3D object which can be rendered.

Object2D

Represents a 2D object which can be rendered.

Shadable

Represents a 3D object that is possible to render with a material that implements the Material trait.

Shadable2D

Represents a 2D object that is possible to render with Materials.