Trait three_d::renderer::object::Object

pub trait Object: Geometry {
    fn render(&self, camera: &Camera, lights: &Lights) -> ThreeDResult<()>;

    fn is_transparent(&self) -> bool;
}

Represents a 3D object which can be rendered.

Required methods

fn render(&self, camera: &Camera, lights: &Lights) -> ThreeDResult<()>

Render the object. Must be called in a render target render function, for example in the callback function of Screen::write. You can use Lights::default() if you know the object does not require lights to be rendered.

fn is_transparent(&self) -> bool

Returns whether or not this object should be considered transparent.