Module three_d::window

Window, event handling and context creation for easy setup.

Can be avoided fully by setting up a window, event handling etc. and creating a Context from a glow OpenGL/WebGL context.

If full control over the window and event handling, but not the context creation, is desired, use a WindowedContext or [HeadlessContext].

Finally, for an easy setup, use Window::new or Window::from_winit_window, the latter will provide full control over the creation of the window.

Structs

FrameInput

Input for rendering (and whatever else needs it) each frame. It includes events that can be used as input to controls. The data should only be used for one frame.

FrameInputGenerator

Use this to generate FrameInput for a new frame with a custom winit window. FrameInput is automatically generated if using the default Window.

FrameOutput

Output from the rendering to the default Window each frame.

SurfaceSettings

Settings controlling the behavior of the surface on where to draw, to present it on the screen.

Window

Default window, context and event handling which uses winit.

WindowSettings

Settings for the default Window.

WindowedContext

A windowed graphics context, ie. a graphics context that is associated with a window. For a graphics context that is not associated with a window, see HeadlessContext.

Enums

HardwareAcceleration

Selects the level of hardware graphics acceleration.

WindowError

Error associated with a window.